CDRIVER-4599 Reducing Warnings - MSVC and MinGW Compilation Warnings (#1229)
* Disable C4756 when compiling with VS 2013 or older

* client-side-encryption-auto-decryption.c: address C4267 size_t conversion warnings

* client-side-encryption-explicit.c: address C4267 size_t conversion warnings

* client-side-encryption-helpers.c: address C4267 size_t conversion warnings

* mcd-azure.c: address -Wmissing-braces warnings

* mcd-azure.c: address C4028 parameter inconsistency warnings

* mongoc-async-cmd.c: address C4018 signedness mismatch warnings

* mongoc-bulk-operation.c: address C4018 signedness mismatch warnings

* mongoc-client-side-encryption.c: address C4018 signedness mismatch warnings

* mongoc-client-side-encryption.c: address C4028 parameter inconsistency warnings

* mongoc-client.c: address -Wmissing-braces warnings

* mongoc-cluster.c: address C4018 signedness mismatch warnings

* mongoc-cluster.c: address C4267 size_t conversion warnings

* mongoc-collection.c: address -Wmissing-braces warnings

* mongoc-compression.c: address C4267 size_t conversion warnings

* mongoc-crypt.c: address C4018 signedness mismatch warnings

* mongoc-crypt.c: address C4244 narrowing conversion warnings

* mongoc-cyrus.c: address 4267 size_t conversion warnings

* mongoc-gridfs-bucket-file.c: address C4267 size_t conversion warnings

* mongoc-gridfs-file.c: address C4018 signedness mismatch warnings

* mongoc-gridfs-file.c: address C4244 narrowing conversion warnings

* mongoc-gridfs-file.c: address scan-build uninitialized value warnings

* mongoc-handshake.c: address C4018 signedness mismatch warnings

* mongoc-http.c: address C4244 narrowing conversion warnings

* mongoc-http.c: address C4267 size_t conversion warnings

* mongoc-log.c: address C4018 signedness mismatch warnings

* mongoc-opts-helpers.c: address C4267 size_t conversion warnings

* mongoc-rpc.c: address C4267 size_t conversion warnings

* mongoc-scram.c: address C4018 signedness mismatch warnings

* mongoc-scram.c: address C4267 size_t conversion warnings

* mongoc-server-monitor.c: address -Wmissing-braces warnings

* mongoc-server-monitor.c: address C4244 narrowing conversion warnings

* mongoc-server-monitor.c: address C4267 size_t conversion warnings

* mongoc-set.c: address C4018 signedness mismatch warnings

* mongoc-set.c: address C4244 narrowing conversion warnings

* mongoc-set.c: address C4267 size_t conversion warnings

* mongoc-shared.c: address C4028 parameter inconsistency warnings

* mongoc-socket.c: address C4018 signedness mismatch warnings

* mongoc-socket.c: address C4267 size_t conversion warnings

* mongoc-stream-file.c: address C4267 size_t conversion warnings

* mongoc-stream-socket.c: address C4018 signedness mismatch warnings

* mongoc-stream-tls-openssl.c: address C4018 signedness mismatch warnings

* mongoc-stream-tls-openssl.c: address C4244 narrowing conversion warnings

* mongoc-stream-tls-openssl.c: address C4267 size_t conversion warnings

* mongoc-stream-tls-secure-channel.c: address C4018 signedness mismatch warnings

* mongoc-stream-tls-secure-channel.c: address C4244 narrowing conversion warnings

* mongoc-stream-tls.c: address C4244 narrowing conversion warnings

* mongoc-stream.c: address C4018 signedness mismatch warnings

* mongoc-topology-background-monitoring.c: address C4018 signedness mismatch warnings

* mongoc-topology-background-monitoring.c: address C4267 size_t conversion warnings

* mongoc-topology-description.c: address C4018 signedness mismatch warnings

* mongoc-topology-description.c: address C4267 size_t conversion warnings

* mongoc-topology-scanner.c: address C4267 size_t conversion warnings

* mongoc-topology.c: address C4028 parameter inconsistency warnings

* mongoc-topology.c: address C4244 narrowing conversion warnings

* mongoc-util.c: address C4267 size_t conversion warnings

* mongoc-util.c: address scan build null pointer warnings

* mongoc-write-command-legacy.c: address C4018 signedness mismatch warnings

* mongoc-write-command.c: address -Wmissing-braces warnings

* mongoc-write-command.c: address C4018 signedness mismatch warnings

* op-msg.def: address C4267 size_t conversion warnings

* service-gcp.c: address -Wmissing-braces warnings

* TestSuite.c: address C4018 signedness mismatch warnings

* entity-map.c: address -Wstrict-prototypes warnings

* json-test-operations.c: address C4018 signedness mismatch warnings

* json-test-operations.c: address C4244 narrowing conversion warnings

* json-test.c: address C4018 signedness mismatch warnings

* mock-rs.c: address C4018 signedness mismatch warnings

* mock-server.c: address C4267 size_t conversion warnings

* operation.c: address C4018 signedness mismatch warnings

* operation.c: address C4244 narrowing conversion warnings

* operation.c: address C4267 size_t conversion warnings

* request.c: address C4018 signedness mismatch warnings

* runner.c: address C4018 signedness mismatch warnings

* test-bson.c: address C4018 signedness mismatch warnings

* test-conveniences.c: address C4018 signedness mismatch warnings

* test-happy-eyeballs.c: address C4018 signedness mismatch warnings

* test-libmongoc.c: address -Wunused-parameter warnings

* test-libmongoc.c: suppress C4091 warning generated by DbgHelp.h

* test-mongoc-aws.c: address C4267 size_t conversion warnings

* test-mongoc-bulk.c: address C4267 size_t conversion warnings

* test-mongoc-client-session.c: address C4267 size_t conversion warnings

* test-mongoc-client-side-encryption.c: address C4267 size_t conversion warnings

* test-mongoc-connection-uri.c: address C4244 narrowing conversion warnings

* test-mongoc-dns.c: address C4244 narrowing conversion warnings

* test-mongoc-dns.c: address C4267 size_t conversion warnings

* test-mongoc-gridfs-bucket.c: address C4267 size_t conversion warnings

* test-mongoc-gridfs.c: address C4244 size_t conversion warnings

* test-mongoc-handshake.c: address C4267 size_t conversion warnings

* test-mongoc-interrupt.c: address C4244 narrowing conversion warnings

* test-mongoc-long-namespace.c: address C4267 size_t conversion warnings

* test-mongoc-mongohouse.c: address C4244 narrowing conversion warnings

* test-mongoc-read-write-concern.c: address C4244 narrowing conversion warnings

* test-mongoc-rpc.c: address C4267 size_t conversion warnings

* test-mongoc-sdam.c: address -Wmissing-braces warnings

* test-mongoc-sdam.c: address C4018 signedness mismatch warnings

* test-mongoc-sdam.c: address C4267 size_t conversion warnings

* test-mongoc-socket.c: address C4244 narrowing conversion warnings

* test-mongoc-topology.c: address C4244 narrowing conversion warnings

* test-mongoc-uri.c: address C4244 narrowing conversion warnings

* test-mongoc-util.c: address C4267 size_t conversion warnings

* CDRIVER-4263: address C4996 deprecated function warnings

+++ src/libmongoc/src/mongoc/mongoc-set-private.h
@@ -64,20 +64,22 @@ mongoc_set_get_const (const mongoc_set_t *set, uint32_t id)
 }
 
 void *
-mongoc_set_get_item (mongoc_set_t *set, int idx);
+mongoc_set_get_item (mongoc_set_t *set, size_t idx);
 
 static BSON_INLINE const void *
-mongoc_set_get_item_const (const mongoc_set_t *set, int idx)
+mongoc_set_get_item_const (const mongoc_set_t *set, size_t idx)
 {
    return mongoc_set_get_item ((mongoc_set_t *) set, idx);
 }
 
 void *
-mongoc_set_get_item_and_id (mongoc_set_t *set, int idx, uint32_t *id /* OUT */);
+mongoc_set_get_item_and_id (mongoc_set_t *set,
+                            size_t idx,
+                            uint32_t *id /* OUT */);
 
 static BSON_INLINE const void *
 mongoc_set_get_item_and_id_const (const mongoc_set_t *set,
-                                  int idx,
+                                  size_t idx,
                                   uint32_t *id)
 {
    return mongoc_set_get_item_and_id ((mongoc_set_t *) set, idx, id);
